# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /v1/accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List connected accounts
      description: Returns all social accounts connected by the current user.
      responses:
        "200":
          description: Connected accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountsCollection"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Connect an account
      description: Creates a connected account record for the current user. OAuth exchange is not handled by this endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Connected account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/accounts/youtube/connect:
    get:
      operationId: connectYouTubeAccount
      tags: [accounts]
      summary: Start YouTube OAuth
      description: Redirects the authenticated user to Google's YouTube OAuth consent screen.
      responses:
        "302":
          description: Redirect to Google OAuth
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/accounts/youtube/callback:
    get:
      operationId: handleYouTubeAccountCallback
      tags: [accounts]
      summary: Complete YouTube OAuth
      description: Handles Google's OAuth callback and stores YouTube tokens for the account.
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect after account connection
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/accounts/tiktok/connect:
    get:
      operationId: connectTikTokAccount
      tags: [accounts]
      summary: Start TikTok OAuth
      description: Redirects the authenticated user to TikTok's OAuth consent screen.
      responses:
        "302":
          description: Redirect to TikTok OAuth
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/accounts/tiktok/callback:
    get:
      operationId: handleTikTokAccountCallback
      tags: [accounts]
      summary: Complete TikTok OAuth
      description: Handles TikTok's OAuth callback and stores TikTok tokens for the account.
      security: []
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: scopes
          in: query
          required: false
          schema:
            type: string
        - name: error
          in: query
          required: false
          schema:
            type: string
        - name: error_description
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect after account connection
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/accounts/meta/connect:
    get:
      operationId: connectMetaAccount
      tags: [accounts]
      summary: Start Meta OAuth
      description: Redirects the authenticated user to Meta's OAuth consent screen for Facebook Pages and Instagram professional accounts.
      responses:
        "302":
          description: Redirect to Meta OAuth
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/accounts/meta/callback:
    get:
      operationId: handleMetaAccountCallback
      tags: [accounts]
      summary: Complete Meta OAuth
      description: Handles Meta's OAuth callback and stores Facebook Page and Instagram account tokens.
      security: []
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: error
          in: query
          required: false
          schema:
            type: string
        - name: error_code
          in: query
          required: false
          schema:
            type: string
        - name: error_message
          in: query
          required: false
          schema:
            type: string
        - name: error_description
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect after account connection
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get connected account
      description: Returns a single connected account owned by the current user.
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "200":
          description: Connected account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update connected account
      description: Updates the display name or metadata of a connected account owned by the current user.
      parameters:
        - $ref: "#/components/parameters/AccountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Disconnect an account
      description: Deletes a connected account owned by the current user.
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "204":
          description: Account disconnected
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT access token.
  parameters:
    AccountId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    AccountPlatform:
      type: string
      enum:
        - facebook
        - instagram
        - tiktok
        - youtube
        - snapchat
    Account:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        platform:
          $ref: "#/components/schemas/AccountPlatform"
        accountName:
          type: string
          minLength: 1
          maxLength: 255
        tokenExpiry:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        userId:
          type: string
          format: uuid
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - platform
        - accountName
        - tokenExpiry
        - userId
        - metadata
        - createdAt
    AccountsCollection:
      type: object
      additionalProperties: false
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Account"
      required:
        - data
    CreateAccountRequest:
      type: object
      additionalProperties: false
      properties:
        platform:
          $ref: "#/components/schemas/AccountPlatform"
        accountName:
          type: string
          minLength: 1
          maxLength: 255
        accessToken:
          type: string
          minLength: 1
        refreshToken:
          oneOf:
            - type: string
              minLength: 1
            - type: "null"
        tokenExpiry:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        metadata:
          type: object
          additionalProperties: true
      required:
        - platform
        - accountName
        - accessToken
    UpdateAccountRequest:
      type: object
      additionalProperties: false
      properties:
        accountName:
          type: string
          minLength: 1
          maxLength: 255
        metadata:
          type: object
          additionalProperties: true
...
    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
      required:
        - statusCode
        - error
        - message